update install method
