+++ RecDP/tests/cicd/bashrun_unittest_llmutils.sh
@@ -2,7 +2,7 @@
 
 failed_tests=""
 echo "Setup pyrecdp latest package"
-python setup.py sdist && pip install dist/pyrecdp-*.*.*.tar.gz
+pip install -e .[LLM]
 
 # call cmdline tests
 cmdline="python pyrecdp/primitives/llmutils/quality_classifier.py --dataset_path tests/data/llm_data/arxiv_sample_100.jsonl --result_path tests/data/output/qualify_classify"
